refactor: 사이드바 UI 구조 수정
